payment/requirements.txt
fastapi~=0.89.1
starlette~=0.22.0
python-dotenv~=0.21.1
requests~=2.25.1
uvicorn~=0.15.0